Add files via upload
This is the model for Decision Tree for predicting the previous dataset i had used in previous regressions.
This is much better than previous model though progress is not signigicant but even accuracy increase in points matters here.
But Best one is yet to come that is RandomForest regression.
Happy! Learning.
Go! for Data Science.
I am learning too...